Optimized enum usage and added [[maybe_unused]] attribute
In the `CompNodeType.hpp` file, the `comp_NodeType` function has been modified to use the `NodeType` enumeration without the `NodeType::` prefix for each case in the switch statement. This was achieved by adding `using enum NodeType;` at the beginning of the function.

In the `Transpiler.cpp` file, the `[[maybe_unused]]` attribute was added to the `nullptrNode` variable to indicate that it may not be used, thereby avoiding potential compiler warnings.