Update ci.yml and build_and_test_coverage.sh
Refactor the ci.yml and build_and_test_coverage.sh scripts to improve code coverage reporting. Updated the gcovr command in both files to include the root directory and use the gcovr.cfg configuration file. Also added the out/ directory to the .gitignore file to ensure it is not ignored.