Whitespace management refactor in Tokenizer
The handleWhiteSpaceSingle() function was added to handle a single whitespace character, replacing handleWhiteSpace() in two places in Tokenizer.cpp. The handleWhiteSpace() function was modified to handle multiple consecutive whitespace characters using a while loop. The changes were made in the Tokenizer.hpp and Tokenizer.cpp files within the vnd namespace.

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>